Finish integration of test vectors and fix more bugs
Fixes a bug that entirely disabled using proposal references
in commits. Fixes calculation of proposal references to use
AuthenticatedContent<Proposal> instead of the raw Proposal.

Fixes enum encoded width of ResumptionPskUsage to be one byte
instead of two.

Relax deserialization of the ReInit extensions field, as it
previously required the extension types to be valid for the
GroupContext. Unfortunately, the test vectors use an illegal
extension type (RatchetTree) in there, which made tests fail.
If this is fixed in the test vectors, this change can be reverted.

Signed-off-by: Johannes Leupold <johannes.leupold@posteo.de>