Implement a rough sketch for the high-level Re-init API
Processing handshake messages now returns a dedicated
result type, if successful. It may indicate that

a. a proposal was processed
b. a commit was processed (that added no new members)
c. a commit was processed that added new members
d. a commit was processed that triggered a re-init

In case c., the result includes a list of welcome
messages to send, plus the key packages they are
to be sent to. Note that only the committer receives
this result, as other members may not be able to create
the welcome messages.

In case d., the result contains the SuspendedGroupClient
instance that can be used to resume the group (i.e start
a new group and create welcome messages for all members).
Note that this transition to a different client state
isn't supported on the MlsClient level so far.

Removes the very outdated and incomplete service modules

Signed-off-by: Johannes Leupold <johannes.leupold@posteo.de>